Ammazzato bug e adesso il test che dava errori è stato sistemato.

+++ Nella oscura e pronda caverna/Nella oscura e pronda caverna/Entita.cpp
@@ -59,7 +59,7 @@ Danno Entita::attacca()
 			danno.magnifica(attributi.getDestrezza() / 4);
 		else										//arma grande no
 			danno.magnifica(attributi.getForza() / 4);
-		srand(time(nullptr));
+		srand((unsigned int)time(nullptr));
 		//FIXMe return sempre = 1
 		double random;
 		random = rand() % 6;
@@ -97,7 +97,7 @@ bool Entita::addInventario(std::list<std::shared_ptr<Oggetto>> oggettiAggiunti)
 void Entita::equip(int posizioneFrom, int posiozioneTo) {
 	if (posiozioneTo > 40) //troppo in là nell'inventario
 		return;
-	if (posizioneFrom<0 || posizioneFrom>inventario.size())
+	if (posizioneFrom<0 ||(unsigned int) posizioneFrom>inventario.size())
 		return; //non è dentro
 	else {
 		auto it = inventario.begin();
@@ -110,7 +110,7 @@ void Entita::equip(int posizioneFrom, int posiozioneTo) {
 }
 
 void Entita::unequip(int posisioneFrom) {
-	if (posisioneFrom<0 || posisioneFrom>inventario.size())
+	if (posisioneFrom<0 ||(unsigned int) posisioneFrom>inventario.size())
 		return; //non è dentro
 	else {
 		auto moving = equipaggiamento.at(posisioneFrom);
